refactor: move state types to own file
