+++ src/app/state/reducer.ts
@@ -5,21 +5,7 @@ import {
   getReverseCypher,
   type Cypher,
 } from "@/utils/cypher";
-
-export type State = {
-  cypher: Cypher;
-  answerCypher: Cypher;
-  quote: string;
-  author: string;
-  encryptedQuote: string;
-  letterFrequency: Map<string, number>;
-  currentLetter: string;
-  loading?: boolean;
-  win?: boolean;
-  completeWithError?: boolean;
-  msElapsed: number;
-  lastTick: number;
-};
+import { ActionType, type Action, type State } from "./types";
 
 export const initialState: State = {
   cypher: generateCypher(),
@@ -35,31 +21,6 @@ export const initialState: State = {
   msElapsed: 0,
   lastTick: Date.now(),
 };
-
-export enum ActionType {
-  Clear = "clear",
-  NewGame = "newGame",
-  GiveUp = "giveUp",
-  Loading = "loading",
-  SetAnswer = "setAnswer",
-  SetCurrentLetter = "setCurrentLetter",
-  Tick = "tick",
-}
-
-export type Action =
-  | { type: ActionType.Clear }
-  | {
-      type: ActionType.NewGame;
-      payload: { quote: string; author: string };
-    }
-  | { type: ActionType.GiveUp }
-  | { type: ActionType.Loading }
-  | {
-      type: ActionType.SetAnswer;
-      payload: { encoded: string; decoded: string };
-    }
-  | { type: ActionType.SetCurrentLetter; payload: string }
-  | { type: ActionType.Tick; payload: number };
 
 export default function reducer(state: State, action: Action): State {
   switch (action.type) {
